+++ src/RestApi.js
@@ -1,0 +1,11 @@
+const PRODUCTION_BASE_PATH = 'https://docusign.net/restapi';
+const DEMO_BASE_PATH = 'https://demo.docusign.net/restapi';
+const STAGE_BASE_PATH = 'https://stage.docusign.net/restapi';
+
+module.exports = {
+  BasePath: {
+    PRODUCTION: PRODUCTION_BASE_PATH,
+    DEMO: DEMO_BASE_PATH,
+    STAGE: STAGE_BASE_PATH
+  }
+};